menu validation

+++ database/migrations/2020_11_06_134751_create_restaurent_management_table.php
@@ -53,9 +53,9 @@ class CreateRestaurentManagementTable extends Migration
             $table->decimal('commission', 8, 2)->nullable();
             $table->decimal('tax', 8, 2)->nullable();
             $table->decimal('flat_delivery_charge', 8, 2)->nullable();
-            $table->tinyInteger('on_km')->nullable();
+            $table->integer('on_km')->nullable();
             $table->decimal('after_flat_delivery_charge', 8, 2)->nullable();
-            $table->tinyInteger('listing_order')->nullable();
+            $table->integer('listing_order')->nullable();
             $table->tinyInteger('visibility')->default('0');
             $table->timestamp('deleted_at', 0)->nullable();
             $table->timestamps();
@@ -71,7 +71,7 @@ class CreateRestaurentManagementTable extends Migration
             $table->foreign('service_catagory_id')->references('id')->on('service_catagories')->onDelete('cascade')->onUpdate('cascade');
             //rest attributes
             $table->string('name');
-            $table->tinyInteger('listing_order')->nullable();
+            $table->integer('listing_order')->nullable();
             $table->tinyInteger('visibility')->default('0');
             $table->timestamp('deleted_at', 0)->nullable();
             $table->timestamps();
@@ -114,7 +114,7 @@ class CreateRestaurentManagementTable extends Migration
             $table->string('name');
             $table->string('picture')->nullable();
             $table->text('about')->nullable();
-            $table->tinyInteger('product_variant_id')->nullable();
+            $table->integer('product_variant_id')->nullable();
             $table->json('product_add_on_id')->nullable();
             $table->decimal('price', 8, 2)->nullable();
             $table->decimal('discount', 8, 2)->nullable();
@@ -136,7 +136,7 @@ class CreateRestaurentManagementTable extends Migration
             $table->foreign('service_catagory_id')->references('id')->on('service_catagories')->onDelete('cascade')->onUpdate('cascade');
             //rest attributes
             $table->string('name');
-            $table->tinyInteger('listing_order')->nullable();
+            $table->integer('listing_order')->nullable();
             $table->tinyInteger('visibility')->default('0');
             $table->timestamp('deleted_at', 0)->nullable();
             $table->timestamps();
